Add wiremock dependency
We can import those mappings directly into the tests.
A bit of discussion is possible here - in a real project I'd definitely do this, or add a contract dependency if we want to get fancy.

But in this case there also is code in the project providing the stubs - and I haven't looked closely enough at it to guarantee merely importing the json stubs is going to produce the same behaviour.
For sake of development speed I may not actually end up doing that. We'll see.