Put jupiter back in.
Tried getting RestAssured and Wiremock to work for a while. Time to give up on that. This thing needs to work, first.
The silly thing there is that I have that setup working in other projects, it's nothing new, but Spring boot refuses to play ball and
do it's configuration resolution and bean wiring thing properly.

Maybe it's just something in the configuration, but what ..?

Maybe if I start implementing the application itself we're going to get this project into a state where that problem just goes away on it's own.