Stop referring to card references in the PowerOfAttorneyDTO
Ideally we have just one endpoint for each view, which an application can call to get all of the information it needs to implement it's custom view of the data.
That way we reduce roundtrips and make things easier on the app developer, whom obviously does not care what applications are involved in creating the custom view.
He just wants his data..